Added in solution for problem 3
Currently, problem_3 will contain a function to get the total amount of patterns and store into a dictionary. length can be defined.
Also contains a function to compare a pattern to a series of sequences to get the total number of mismatches.
Main function includes going through all possible patterns and comparing them and reporting best pattern with lowest amount of mismatches.

Still need to get location of where the pattern was found and format the printing a bit better.